+++ lib/predicate/nodes/in.rb
@@ -6,32 +6,33 @@ class Predicate
       80
     end
 
-    def identifier
+    def left
       self[1]
     end
+    alias :identifier :left
 
-    def values
+    def right
       self[2]
     end
 
     def &(other)
-      case other
-      when In
-        fv = free_variables
-        if fv.size == 1 && fv == other.free_variables
-          intersection = values & other.values
-          if intersection.empty?
-            Factory.contradiction
-          elsif intersection.size == 1
-            Factory.eq(fv.first, [:literal, intersection.first])
-          else
-            Factory.in(fv.first, intersection)
-          end
-        else
-          super
-        end
+      # we only optimize with another In
+      return super unless other.is_a?(In)
+
+      # we only optimize is same free variables
+      fv = free_variables
+      return super unless fv.size == 1 && fv == other.free_variables
+
+      # we only optimize if both right terms are literals
+      return super unless right.literal? and other.right.literal?
+
+      intersection = right.value & other.right.value
+      if intersection.empty?
+        Factory.contradiction
+      elsif intersection.size == 1
+        Factory.eq(fv.first, [:literal, intersection.first])
       else
-        super
+        Factory.in(fv.first, intersection)
       end
     end
 
@@ -40,11 +41,19 @@ class Predicate
     end
 
     def constant_variables
-      values.size == 1 ? free_variables : []
+      if right.literal? and right.value.size == 1
+        free_variables
+      else
+        []
+      end
     end
 
     def constants
-      values.size == 1 ? { identifier.name => values.first } : {}
+      if right.literal? and right.value.size == 1
+        { identifier.name => right.value.first }
+      else
+        {}
+      end
     end
 
     def dyadic_priority
@@ -52,6 +61,7 @@ class Predicate
     end
 
     def evaluate(tuple)
+      values = right.evaluate(tuple)
       values.include?(identifier.evaluate(tuple))
     end